Bug 5110 (#527): changed type for memory value to deal with >2TB hosts
(cherry picked from commit 248b75094550b18ab26ba9f9a1c059a9ff8d1c89)

+++ src/vm/VirtualMachine.cc
@@ -736,8 +736,8 @@ int VirtualMachine::insert(SqlDB * db, string& error_str)
     string prefix;
 
     string value;
-    int    ivalue;
-    float  fvalue;
+    long int ivalue;
+    float fvalue;
     set<int> cluster_ids;
     vector<Template *> quotas;
 
@@ -1752,7 +1752,7 @@ void VirtualMachine::get_requirements (int& cpu, int& memory, int& disk,
 /* -------------------------------------------------------------------------- */
 
 int VirtualMachine::check_resize (
-        float cpu, int memory, int vcpu, string& error_str)
+        float cpu, long int memory, int vcpu, string& error_str)
 {
     if (cpu < 0)
     {
@@ -1778,7 +1778,7 @@ int VirtualMachine::check_resize (
 /* -------------------------------------------------------------------------- */
 /* -------------------------------------------------------------------------- */
 
-int VirtualMachine::resize(float cpu, int memory, int vcpu, string& error_str)
+int VirtualMachine::resize(float cpu, long int memory, int vcpu, string& error_str)
 {
     ostringstream oss;
